Drawing shape, displaying it on multiple platforms. Probably wrong appr.
Used the ShapeDrawer library to try and dynamically draw the outline of the puzzle piece. Before I go forward with trying to color these pieces with infill algorithms or through even more complicated approaches, I will try using a TenPatch textures approach to dynamically draw and stretch them. It might be easier and less math-heavy to implement and maintain. If it works at all.